Engine: csci gui messages uses intptr_t insted of long

+++ Engine/gui/guidialog.cpp
@@ -189,11 +189,11 @@ int savegamedialog()
 
   lpTemp = nullptr;
   if (numsaves > 0)
-    CSCISendControlMessage(ctrllist, CLB_GETTEXT, 0, (long)&buffer2[0]);
+    CSCISendControlMessage(ctrllist, CLB_GETTEXT, 0, (intptr_t)&buffer2[0]);
   else
     buffer2[0] = 0;
 
-  CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (long)&buffer2[0]);
+  CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (intptr_t)&buffer2[0]);
 
   int toret = -1;
   while (1) {
@@ -201,10 +201,10 @@ int savegamedialog()
     if (mes.code == CM_COMMAND) {
       if (mes.id == ctrlok) {
         int cursell = CSCISendControlMessage(ctrllist, CLB_GETCURSEL, 0, 0);
-        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, (long)&buffer2[0]);
+        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, (intptr_t)&buffer2[0]);
 
         if (numsaves > 0)
-          CSCISendControlMessage(ctrllist, CLB_GETTEXT, cursell, (long)&bufTemp[0]);
+          CSCISendControlMessage(ctrllist, CLB_GETTEXT, cursell, (intptr_t)&bufTemp[0]);
         else
           strcpy(bufTemp, "_NOSAVEGAMENAME");
 
@@ -228,7 +228,7 @@ int savegamedialog()
             CSCIWaitMessage(&cmes);
           } while (cmes.code != CM_COMMAND);
 
-          CSCISendControlMessage(txt1, CTB_GETTEXT, 0, (long)&buffer2[0]);
+          CSCISendControlMessage(txt1, CTB_GETTEXT, 0, (intptr_t)&buffer2[0]);
           CSCIDeleteControl(btnCancel);
           CSCIDeleteControl(btnOk);
           CSCIDeleteControl(txt1);
@@ -279,8 +279,8 @@ int savegamedialog()
     } else if (mes.code == CM_SELCHANGE) {
       int cursel = CSCISendControlMessage(ctrllist, CLB_GETCURSEL, 0, 0);
       if (cursel >= 0) {
-        CSCISendControlMessage(ctrllist, CLB_GETTEXT, cursel, (long)&buffer2[0]);
-        CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (long)&buffer2[0]);
+        CSCISendControlMessage(ctrllist, CLB_GETTEXT, cursel, (intptr_t)&buffer2[0]);
+        CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (intptr_t)&buffer2[0]);
       }
     }
   }
@@ -304,7 +304,7 @@ void preparesavegamelist(int ctrllist)
   // fill in the list box and global savegameindex[] array for backward compatibilty
   for (numsaves = 0; (size_t)numsaves < saves.size(); ++numsaves)
   {
-      CSCISendControlMessage(ctrllist, CLB_ADDITEM, 0, (long)saves[numsaves].Description.GetCStr());
+      CSCISendControlMessage(ctrllist, CLB_ADDITEM, 0, (intptr_t)saves[numsaves].Description.GetCStr());
       filenumbers[numsaves] = saves[numsaves].Slot;
       filedates[numsaves] = (long int)saves[numsaves].FileTime;
   }
@@ -339,7 +339,7 @@ void enterstringwindow(const char *prompttext, char *stouse)
       if (mes.id == ctrlcancel)
         buffer2[0] = 0;
       else
-        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, (long)&buffer2[0]);
+        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, (intptr_t)&buffer2[0]);
       break;
     }
   }
@@ -382,7 +382,7 @@ int roomSelectorWindow(int currentRoom, int numRooms,
   for (int aa = 0; aa < numRooms; aa++)
   {
     snprintf(buff, sizeof(buff), "%3d %s", roomNumbers[aa], roomNames[aa].GetCStr());
-    CSCISendControlMessage(ctrllist, CLB_ADDITEM, 0, (long)&buff[0]);
+    CSCISendControlMessage(ctrllist, CLB_ADDITEM, 0, (intptr_t)&buff[0]);
     if (roomNumbers[aa] == currentRoom)
     {
       CSCISendControlMessage(ctrllist, CLB_SETCURSEL, aa, 0);
@@ -397,7 +397,7 @@ int roomSelectorWindow(int currentRoom, int numRooms,
   buffer2[0] = 0;
 
   int ctrltbox = CSCICreateControl(CNT_TEXTBOX, 10, 29, 120, 0, nullptr);
-  CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (long)&buffer2[0]);
+  CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (intptr_t)&buffer2[0]);
 
   int toret = -1;
   while (1) {
@@ -406,7 +406,7 @@ int roomSelectorWindow(int currentRoom, int numRooms,
     {
       if (mes.id == ctrlok) 
       {
-        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, (long)&buffer2[0]);
+        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, (intptr_t)&buffer2[0]);
         if (isdigit(buffer2[0]))
         {
           toret = atoi(buffer2);
@@ -423,7 +423,7 @@ int roomSelectorWindow(int currentRoom, int numRooms,
       if (cursel >= 0) 
       {
         snprintf(buffer2, sizeof(buffer2), "%d", roomNumbers[cursel]);
-        CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (long)&buffer2[0]);
+        CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (intptr_t)&buffer2[0]);
       }
     }
   }
