+++ Engine/gui/newcontrol.h
@@ -29,7 +29,7 @@ struct NewControl
   char needredraw;
   virtual void draw(Common::Bitmap *ds) = 0;
   virtual int pressedon(int mx, int my) = 0;
-  virtual int processmessage(int, int, long) = 0;
+  virtual int processmessage(int, int, intptr_t) = 0;
 
   NewControl(int xx, int yy, int wi, int hi);
   NewControl();
